requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argcomplete==3.1.4
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
boto3==1.33.7
botocore==1.33.7
certifi==2023.11.17
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
Django==5.0
durationpy==0.5
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonfy==0.4
jsonpatch==1.33
jsonpointer==2.4
kappa==0.6.0
langchain==0.0.345
langchain-core==0.0.9
langsmith==0.0.69
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==0.28.0
packaging==23.2
placebo==0.9.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.3
zappa==0.58.0
